+++ src/test/variables/input.spec.ts
@@ -24,17 +24,34 @@ describe('variables.input', () => {
 
     expect(spy).toHaveBeenCalledWith('Auth Token?', 'jwt.something', true);
   });
-  it('input-askonce', async () => {
-    initFileProvider();
-    userInteractionProvider.showInputPrompt = async msg => msg;
-    const spy = jest.spyOn(userInteractionProvider, 'showInputPrompt');
+  describe('input-askonce', () => {
+    it('should only aks once for jwt.something', async () => {
+      initFileProvider();
+      userInteractionProvider.showInputPrompt = async msg => msg;
+      const spy = jest.spyOn(userInteractionProvider, 'showInputPrompt');
 
-    await sendHttp(`
+      await sendHttp(`
 @test={{ $input-askonce Auth Token? $value: jwt.something }}
 @test2={{ $input-askonce Auth Token? $value: jwt.something }}
     `);
-    expect(spy).toHaveBeenCalledTimes(1);
-    expect(spy).toHaveBeenCalledWith('Auth Token?', 'jwt.something', false);
+      expect(spy).toHaveBeenCalledTimes(1);
+      expect(spy).toHaveBeenCalledWith('Auth Token?', 'jwt.something', false);
+    });
+    it('should not ask for existing variables', async () => {
+      initFileProvider();
+      userInteractionProvider.showInputPrompt = async msg => msg;
+      const spy = jest.spyOn(userInteractionProvider, 'showInputPrompt');
+
+      await sendHttp(
+        `
+@test={{ $input-askonce token $value: jwt.something }}
+    `,
+        {
+          token: 'something',
+        }
+      );
+      expect(spy).toHaveBeenCalledTimes(0);
+    });
   });
   it('password-askonce', async () => {
     initFileProvider();